requirements.txt
asgiref==3.8.1
certifi==2024.2.2
distlib==0.3.8
Django==5.0.3
djangorestframework==3.15.1
filelock==3.14.0
mysqlclient==2.2.4
pipenv==2023.12.1
platformdirs==4.2.1
sqlparse==0.4.4
tzdata==2024.1
virtualenv==20.26.1
